Break cycle dependency with `pkg/storage/chunk` with `pkg/storage/config`
